Makes Ray Tune optional and fixes submitit checkpointing bug (#272)
* makes ray tuneoptional

* Ignore flake8 errors
