+++ env.common.yml
@@ -21,7 +21,6 @@ dependencies:
   - pip:
     - demjson
     - Pillow
-    - ray[tune]==1.1.0
     - git+https://github.com/rusty1s/pytorch_geometric.git@4ea63d3
     - wandb
     - lmdb==1.1.1
